+++ packages/release/CHANGELOG.md
@@ -7,6 +7,13 @@ and this package adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2026-08-12
+
+### Added
+
+- Isolated Next.js compilation proof for explicitly configured framework
+  export subpaths.
+
 ## [0.1.1] - 2026-08-11
 
 ### Added